Add files via upload
This project is designed to showcase innovations in smart mobility and electric vehicle (EV) technologies. Developed as part of a project for the Center for Smart Mobility and Machines, it highlights key features such as autonomous driving systems, EV infrastructure, and cutting-edge research in smart vehicle technologies. The site serves as a hub for information, resources, and updates on the latest advancements in the field of smart transportation.